requirements-gpu.txt
PyYAML==5.2
fastprogress==0.1.22
tensorflow_gpu==1.15.0
ktrain==0.7.3
